Changed character XP to be a long instead of int.
I doubt anyone will end up with a character with enough xp to hit the limit, but better safe than sorry.

+++ app/src/main/java/ca/secondlifestory/activities/character/CharacterDetailFragment.java
@@ -216,7 +216,7 @@ public class CharacterDetailFragment extends BaseFragment {
                     name.setText(mItem.getName());
                     race.setText(mItem.getRace().getName());
                     characterClass.setText(mItem.getCharacterClass().getName());
-                    totalXp.setText(Integer.toString(mItem.getExperience()));
+                    totalXp.setText(Long.toString(mItem.getExperience()));
                     status.setText(mItem.isLiving() ? "Alive" : "Dead");
                     description.setText(mItem.getDetails());
                 } else {
